# resources/swagger.yaml
openapi: 3.0.0
info:
  title: Pet Daycare Management API
  description: API Rest para acompanhamento e controle da creche de pets
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3001
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /pets:
    get:
      tags:
        - Pets
      summary: Obter todos os pets (apenas trabalhadores)
      operationId: getAllPets
      responses:
        '200':
          description: Pets recuperados com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pet'
        '401':
          description: Token não fornecido ou inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '500':
          description: Erro ao recuperar pets
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string

    post:
      tags:
        - Pets
      summary: Registrar novo pet (apenas trabalhadores)
      operationId: createPet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Bisteca"
                breed:
                  type: string
                  enum: [vira-lata, salsichinha]
                  example: "vira-lata"
                age:
                  type: integer
                  example: 12
                weight:
                  type: number
                  example: 22
                tutorId:
                  type: integer
                  example: 1
              required:
                - name
                - breed
                - age
                - weight
                - tutorId
      responses:
        '201':
          description: Pet criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Pet'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '401':
          description: Token não fornecido ou inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '500':
          description: Erro ao criar pet
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
  /pets/{petId}:
    get:
      tags:
        - Pets
      summary: Obter dados específicos de um pet
      operationId: getPetById
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: name
          in: query
          required: false
          schema:
            type: string
            example: "Bisteca"
      responses:
        '200':
          description: Pet recuperado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Pet'
        '401':
          description: Token não fornecido ou inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '404':
          description: Pet não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '500':
          description: Erro ao recuperar pet
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string

    put:
      tags:
        - Pets
      summary: Atualizar dados de um pet (apenas trabalhadores)
      operationId: updatePet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                breed:
                  type: string
                age:
                  type: integer
                weight:
                  type: number
      responses:
        '200':
          description: Pet atualizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Pet'
        '401':
          description: Token não fornecido ou inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '404':
          description: Pet não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '500':
          description: Erro ao atualizar pet
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string

    delete:
      tags:
        - Pets
      summary: Deletar um pet (apenas trabalhadores)
      operationId: deletePet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Pet deletado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
        '401':
          description: Token não fornecido ou inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '404':
          description: Pet não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '500':
          description: Erro ao deletar pet
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
...
components:
  schemas:
    Pet:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        breed:
          type: string
          enum: [vira-lata, salsichinha]
        age:
          type: integer
        weight:
          type: number
        tutorId:
          type: integer
      required:
        - id
        - name
        - breed
        - age
        - weight
        - tutorId
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT